♻️  router models.py 추출 — 6 dataclass + 7 type alias (P0-P step 3)
변경 이유
- 다른 모든 router 모듈이 dataclass / type alias 에 의존. leaf 모듈을
  먼저 분리해 후속 추출의 dependency root 를 명확히.

주요 변경 사항
- 새 파일 ``engineering_channel_router/models.py`` (218 줄):
  - 6 frozen dataclass (Route/Conversation/Kickoff/Continuation/
    ResearchLoopReport/RouteResult)
  - 7 Callable alias (Send/ExtractPrompt/Conversation/Intake/
    ThreadKickoff/ThreadContinuation/ResearchLoop)
- ``_legacy.py`` 의 dataclass 블록 (162 줄) 제거 → re-export 만 남김.
- ``__init__.py`` facade 가 ``.models`` 에서 직접 import — 외부 caller
  의 type hint resolution 이 ``_legacy`` 의존을 거치지 않게.
- ``EngineeringRouteContext.from_env`` / ``.configured`` 는 utils
  helper 가 필요 → lazy ``from ._legacy import`` (step 4 에서 utils
  분리되면 ``from .utils`` 로 정리 예정).

비고
- 4672 PASS + 1 skipped 회귀 유지.
